refactor: componentのsharedを修正
